+++ ProjectScion/Light.h
@@ -8,21 +8,21 @@
 
 namespace LIGHT_CONSTANTS
 {
-	static const int NUM_SEGMENTS = 32;
-	static const float SEGMENT_ANGLE = 2 * M_PI / NUM_SEGMENTS;
+    static const int NUM_SEGMENTS = 32;
+    static const float SEGMENT_ANGLE = (float)(2.0 * M_PI / NUM_SEGMENTS);
 };
 
 struct Light
 {
-	sf::Color color;
-	float radius, bleed, linearizeFactor;
+    sf::Color color;
+    float radius, bleed, linearizeFactor;
 
-	Light(sf::Vector2f position, float radius, sf::Color color = sf::Color::White, float bleed = 1.0f, float linearizeFactor = 0.2f);
+    Light(sf::Vector2f position, float radius, sf::Color color = sf::Color::White, float bleed = 1.0f, float linearizeFactor = 0.2f);
 private:
-	sf::Vector2f position;
-	bool dirty;
+    sf::Vector2f position;
+    bool dirty;
 
-	friend class LightManager;
+    friend class LightManager;
 };
 
 #endif
